sync-bot-dir/requirements.txt
protobuf==3.19.0
etcd3
pyyaml
kafka
pykafka
kafka-python
loguru